Rewire remaining pages, move availability-slots server actions
