+++ server/availability-slots/queries.tsx
