+++ server/availability-slots/schema.tsx
